Move already MainWindow independent handlers to MessageHandler
